# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "SightingResponse": {
        "properties": {
          "classification_tier_used": {
            "type": "string"
          },
          "confidence_score": {
            "type": "number"
          },
          "datetime": {
            "type": "string"
          },
          "delayed": {
            "type": "boolean"
          },
          "device_id": {
            "type": "integer"
          },
          "has_image": {
            "type": "boolean"
          },
          "id": {
            "type": "integer"
          },
          "species_id": {
            "type": "integer"
          },
          "weather_conditions": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "classification_tier_used",
          "confidence_score",
          "datetime",
          "delayed",
          "device_id",
          "has_image",
          "id",
          "species_id",
          "weather_conditions"
        ],
        "title": "SightingResponse",
        "type": "object"
      },
      "SightingUploadForm": {
        "properties": {
          "classification_tier_used": {
            "type": "string"
          },
          "common_name": {
            "type": "string"
          },
          "confidence_score": {
            "type": "string"
          },
          "delayed": {
            "default": "false",
            "type": "string"
          },
          "image": {
            "contentMediaType": "application/octet-stream",
            "format": "binary",
            "type": "string"
          },
          "scientific_name": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": [
          "classification_tier_used",
          "common_name",
          "confidence_score",
          "image",
          "scientific_name",
          "timestamp"
        ],
        "title": "SightingUploadForm",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "type": "string"
          }
        },
        "required": [
          "access_token"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "DeviceToken": {
        "description": "Long-lived per-device token issued at device registration. Used by the Pi for POST /sightings, POST /classify, and heartbeats.",
        "scheme": "bearer",
        "type": "http"
      },
      "UserJWT": {
        "bearerFormat": "JWT",
        "description": "User JWT from POST /login.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for the Peck Deck smart bird feeder.\n\nTwo distinct kinds of caller, with separate credentials:\n\n* **Users** (web app) \u2014 `POST /login` returns a JWT. Send it as\n  `Authorization: Bearer <jwt>`. Roles are `owner` and `viewer`; owner-only\n  routes are marked as such.\n* **Devices** (Raspberry Pi feeders) \u2014 authenticate with their long-lived\n  device token, also sent as `Authorization: Bearer <token>`. A device token is\n  **not** accepted on user routes and vice versa.\n\nSighting images are stored in PostgreSQL as `bytea` and served from\n`GET /sightings/{id}/image`, which requires a user JWT.\n\n`GET /health` is liveness (never touches the database); `GET /ready` is\nreadiness (checks the database and migration state, 503 when not ready);\n`GET /meta` describes the instance itself (notably whether it is running in\nread-only demo mode).\n\nWith `DEMO_MODE=1` the instance seeds itself on first boot and rejects every\nuser-authenticated write with 403. Device-token routes (`POST /sightings`,\n`POST /classify`, the heartbeat) and `POST /login` keep working, so the demo\nstays live rather than frozen.",
    "title": "Peck Deck API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/login": {
      "post": {
        "deprecated": false,
        "operationId": "LoginLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Document created, URL follows",
            "headers": {}
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "Validation Exception",
                  "examples": [
                    {
                      "detail": "Bad Request",
                      "extra": {},
                      "status_code": 400
                    }
                  ],
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "extra": {
                      "additionalProperties": {},
                      "type": [
                        "null",
                        "object",
                        "array"
                      ]
                    },
                    "status_code": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "detail",
                    "status_code"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bad request syntax or unsupported method"
          }
        },
        "summary": "Login"
      }
    },
...
    "/sightings": {
      "get": {
        "deprecated": false,
        "operationId": "SightingsListSightings",
        "parameters": [
          {
            "allowEmptyValue": false,
            "allowReserved": false,
            "deprecated": false,
            "in": "query",
            "name": "device_id",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "allowEmptyValue": false,
            "allowReserved": false,
            "deprecated": false,
            "in": "query",
            "name": "species_id",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "allowEmptyValue": false,
            "allowReserved": false,
            "deprecated": false,
            "in": "query",
            "name": "from_date",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "allowEmptyValue": false,
            "allowReserved": false,
            "deprecated": false,
            "in": "query",
            "name": "to_date",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "allowEmptyValue": false,
            "allowReserved": false,
            "deprecated": false,
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "type": "integer"
            }
          },
          {
            "allowEmptyValue": false,
            "allowReserved": false,
            "deprecated": false,
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SightingResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Request fulfilled, document follows",
            "headers": {}
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "Validation Exception",
                  "examples": [
                    {
                      "detail": "Bad Request",
                      "extra": {},
                      "status_code": 400
                    }
                  ],
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "extra": {
                      "additionalProperties": {},
                      "type": [
                        "null",
                        "object",
                        "array"
                      ]
                    },
                    "status_code": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "detail",
                    "status_code"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bad request syntax or unsupported method"
          }
        },
        "summary": "ListSightings"
      },
      "post": {
        "deprecated": false,
        "operationId": "SightingsCreate",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SightingUploadForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SightingResponse"
                }
              }
            },
            "description": "Document created, URL follows",
            "headers": {}
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "Validation Exception",
                  "examples": [
                    {
                      "detail": "Bad Request",
                      "extra": {},
                      "status_code": 400
                    }
                  ],
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "extra": {
                      "additionalProperties": {},
                      "type": [
                        "null",
                        "object",
                        "array"
                      ]
                    },
                    "status_code": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "detail",
                    "status_code"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bad request syntax or unsupported method"
          }
        },
        "summary": "Create"
      }
    },
...
    "/sightings/{sighting_id}/image": {
      "get": {
        "deprecated": false,
        "operationId": "SightingsSightingIdImageGetImage",
        "parameters": [
          {
            "deprecated": false,
            "in": "path",
            "name": "sighting_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Request fulfilled, document follows",
            "headers": {}
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "Validation Exception",
                  "examples": [
                    {
                      "detail": "Bad Request",
                      "extra": {},
                      "status_code": 400
                    }
                  ],
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "extra": {
                      "additionalProperties": {},
                      "type": [
                        "null",
                        "object",
                        "array"
                      ]
                    },
                    "status_code": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "detail",
                    "status_code"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bad request syntax or unsupported method"
          }
        },
        "summary": "GetImage"
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}